requirements.txt
click==8.1.3
defusedxml==0.7.1
Flask==2.2.2
fonttools==4.37.3
fpdf2==2.5.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
PyPDF2==2.10.9
svg.path==6.2
Werkzeug==2.2.2
